+++ src/main/kotlin/com/ingsis/snippets/snippet/routes/SnippetController.kt
@@ -1,7 +1,7 @@
 package com.ingsis.snippets.snippet.routes
 
-import com.ingsis.snippets.snippet.Snippet
 import com.ingsis.snippets.snippet.SnippetService
+import com.ingsis.snippets.snippet.model.Snippet
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
